job-runner/pyproject.toml
[project]
name = "job-runner"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.29",
    "openai>=1.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["job_runner"]
